# perf: improve isBlobLike