--- a/benchmarks/core/is-blob-like.mjs
+++ b/benchmarks/core/is-blob-like.mjs
@@ -1,0 +1,64 @@
+import { bench, group, run } from 'mitata'
+import { isBlobLike } from '../../lib/core/util.js'
+
+const buffer = Buffer.alloc(1)
+
+const blob = new Blob(['asd'], {
+  type: 'application/json'
+})
+
+const file = new File(['asd'], 'file.txt', {
+  type: 'text/plain'
+})
+
+const blobLikeStream = {
+  [Symbol.toStringTag]: 'Blob',
+  stream: () => {}
+}
+
+const fileLikeStream = {
+  stream: () => {},
+  [Symbol.toStringTag]: 'File'
+}
+
+const blobLikeArrayBuffer = {
+  [Symbol.toStringTag]: 'Blob',
+  arrayBuffer: () => {}
+}
+
+const fileLikeArrayBuffer = {
+  [Symbol.toStringTag]: 'File',
+  arrayBuffer: () => {}
+}
+
+group('isBlobLike', () => {
+  bench('blob', () => {
+    return isBlobLike(blob)
+  })
+  bench('file', () => {
+    return isBlobLike(file)
+  })
+  bench('blobLikeStream', () => {
+    return isBlobLike(blobLikeStream)
+  })
+  bench('fileLikeStream', () => {
+    return isBlobLike(fileLikeStream)
+  })
+  bench('fileLikeArrayBuffer', () => {
+    return isBlobLike(fileLikeArrayBuffer)
+  })
+  bench('blobLikeArrayBuffer', () => {
+    return isBlobLike(blobLikeArrayBuffer)
+  })
+  bench('buffer', () => {
+    return isBlobLike(buffer)
+  })
+  bench('null', () => {
+    return isBlobLike(null)
+  })
+  bench('string', () => {
+    return isBlobLike('invalid')
+  })
+})
+
+await run()