@KhafraDev 2024-04-08
I don't mind improving the performance, but we should be looking to remove support for third party web streams/Blob/File/etc.