@Uzlopak 2024-04-08
then we would need only replace isBlobLike with o instanceof Blob everywhere. But this PR is not about removing third party File and Blob support.